backends/makrcave/requirements.txt
# Core FastAPI dependencies
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.32
psycopg2==2.9.9
alembic==1.13.2

# Authentication & Security
python-keycloak==3.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data & Serialization
pydantic==2.8.2
pydantic-settings==2.3.0

# Environment & Configuration
python-dotenv==1.0.1

# Caching & Background Tasks
redis==5.0.7

# Monitoring & Logging
structlog==24.1.0
gunicorn==21.2.0
prometheus-fastapi-instrumentator==6.1.0

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Utilities
python-dateutil==2.9.0.post0
